+++ src/main/java/com/ivan/projectmanager/dto/CommentDTO.java
@@ -1,14 +1,14 @@
 package com.ivan.projectmanager.dto;
 
 public class CommentDTO {
-    private int id;
+    private Long id;
     private String text;
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public CommentDTO setId(int id) {
+    public CommentDTO setId(Long id) {
         this.id = id;
         return this;
     }
