+++ src/main/java/com/ivan/projectmanager/repository/RoleRepository.java
@@ -2,5 +2,5 @@ package com.ivan.projectmanager.repository;
 
 import com.ivan.projectmanager.model.Role;
 
-public interface RoleRepository extends CrudRepository<Role,Integer> {
+public interface RoleRepository extends CrudRepository<Role, Long> {
 }
