+++ src/main/java/com/ivan/projectmanager/repository/TeamRepository.java
@@ -2,5 +2,5 @@ package com.ivan.projectmanager.repository;
 
 import com.ivan.projectmanager.model.Team;
 
-public interface TeamRepository extends CrudRepository<Team,Integer> {
+public interface TeamRepository extends CrudRepository<Team, Long> {
 }
